Add heavy tests config to github CI
