+++ .cargo/config.toml
@@ -1,3 +1,3 @@
 [alias]
-integ-test = ["run", "--package", "temporal-sdk-core", "--example", "integ_runner"]
+integ-test = ["run", "--package", "temporal-sdk-core", "--example", "integ_runner", "--"]
 lint = ["clippy", "--all", "--test", "integ_tests", "--test", "load_tests"]
