+++ sdk-core-protos/Cargo.toml
@@ -15,7 +15,7 @@ history_builders = ["uuid", "rand"]
 
 [dependencies]
 anyhow = "1.0"
-base64 = "0.20"
+base64 = "0.21"
 derive_more = "0.99"
 prost = "0.11"
 prost-types = "0.11"
