+++ sdk/Cargo.toml
@@ -16,7 +16,7 @@ categories = ["development-tools"]
 async-trait = "0.1"
 thiserror = "1.0"
 anyhow = "1.0"
-base64 = "0.20"
+base64 = "0.21"
 crossbeam = "0.8"
 derive_more = "0.99"
 futures = "0.3"
